use security handler instead of express middleware for access token validation.
