+++ lib/units/api/index.js
@@ -7,8 +7,6 @@ var cookieSession = require('cookie-session')
 
 var logger = require('../../util/logger')
 
-var auth = require('./middleware/auth')
-
 module.exports = function(options) {
   var log = logger.createLogger('api')
     , app = express()
@@ -25,15 +23,17 @@ module.exports = function(options) {
       swaggerExpress.register(app);
     })
 
+    // Adding options in request, so that swagger controller
+    // can use it.
+    app.use(function(req, res, next) {
+      req.options = options
+      next()
+    })
+
     // TODO: Remove this once frontend is stateless
     app.use(cookieSession({
       name: options.ssid
     , keys: [options.secret]
-    }))
-
-    app.use(auth({
-      secret: options.secret
-    , authUrl: options.authUrl
     }))
 
     server.listen(options.port)
